Consolidate duplicate coroutine benchmarks
